refactor: no need for BaseTool
